Lazily install and configure elm-test (#34)
The package gets installed only when `make tests` is run for the first
time, so that tests support remains optional.

The task also creates an example test file, whose content is delegated
to `elm-test` via the `init` task. This should ease maintenance
connected with library upgrades.

Note that at this point in time, the generated test file includes a
pending spec, which causes the test runner to exit with error. Such
error triggers a Make failure. To avoid this, make tasks in elm.mk are
now run with the `--ignore-errors` flag.